requirements.txt
python
IMAP
SMTP
psutil
PILLOW
opencv-python
pynput
AppOpener